# service/apigateway/docs/swagger.yaml
definitions:
  errors.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      retryable:
        type: boolean
      status:
        type: string
      trace_id:
        type: string
      type:
        $ref: '#/definitions/errors.ErrorType'
      validations:
        items:
          $ref: '#/definitions/errors.ValidationError'
        type: array
    type: object
  errors.ErrorType:
    enum:
    - INTERNAL
    - NOT_FOUND
    - BAD_REQUEST
    - CONFLICT
    - UNAUTHORIZED
    - FORBIDDEN
    - TIMEOUT
    - UNAVAILABLE
    type: string
    x-enum-varnames:
    - ErrorTypeInternal
    - ErrorTypeNotFound
    - ErrorTypeBadRequest
    - ErrorTypeConflict
    - ErrorTypeUnauthorized
    - ErrorTypeForbidden
    - ErrorTypeTimeout
    - ErrorTypeUnavailable
  errors.ValidationError:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
...
  response.ApiResponsePaginationTransaction:
    properties:
      data:
        description: Array of transaction data
        items:
          $ref: '#/definitions/response.TransactionResponse'
        type: array
      message:
        description: Response message
        type: string
      pagination:
        allOf:
        - $ref: '#/definitions/response.PaginationMeta'
        description: Pagination metadata
      status:
        description: Response status
        type: string
    type: object
...
  response.ApiResponseTransaction:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/response.TransactionResponse'
        description: Single transaction data
      message:
        description: Response message
        type: string
      status:
        description: Response status
        type: string
    type: object
...
  response.ApiResponseTransactions:
    properties:
      data:
        description: Array of transaction data
        items:
          $ref: '#/definitions/response.TransactionResponse'
        type: array
      message:
        description: Response message
        type: string
      status:
        description: Response status
        type: string
    type: object
...
  response.ApiResponsesMerchant:
    properties:
      data:
        description: Array of merchant data
        items:
          $ref: '#/definitions/response.MerchantResponse'
        type: array
      message:
        description: Descriptive message
        type: string
      status:
        description: Response status
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      code:
        description: HTTP status code or application error code
        type: integer
      message:
        description: Human-readable error message
        type: string
      status:
        description: Error status ("error", "fail", etc.)
        type: string
    type: object
...
  response.MerchantResponse:
    properties:
      api_key:
        description: Merchant's API key (typically masked in responses)
        type: string
      created_at:
        description: Timestamp when merchant was created (RFC3339 format)
        type: string
      id:
        description: Unique merchant identifier
        type: integer
      name:
        description: Legal name of the merchant
        type: string
      status:
        description: Current status ("active", "suspended", etc.)
        type: string
      updated_at:
        description: Timestamp when merchant was last updated
        type: string
      user_id:
        description: ID of the user who owns/administers this merchant
        type: integer
    type: object
...
  response.PaginationMeta:
    properties:
      current_page:
        description: Current page number (1-based index)
        type: integer
      page_size:
        description: Number of items per page
        type: integer
      total_pages:
        description: Total number of pages available
        type: integer
      total_records:
        description: Total number of items across all pages
        type: integer
    type: object
...
  response.TransactionResponse:
    properties:
      amount:
        description: Transaction amount
        type: integer
      card_number:
        description: Card number used for the transaction
        type: string
      created_at:
        description: When the record was created
        type: string
      id:
        description: Unique transaction identifier
        type: integer
      merchant_id:
        description: ID of the merchant where transaction occurred
        type: integer
      payment_method:
        description: Payment method used (e.g., "credit", "debit")
        type: string
      transaction_no:
        description: Unique transaction reference number
        type: string
      transaction_time:
        description: Timestamp of the transaction
        type: string
      updated_at:
        description: When the record was last updated
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/merchant/stats/transactions:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponsesMerchant'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get all merchant transactions
      tags:
      - Stats Merchant
...
  /api/transaction-query:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all transactions
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of transactions
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationTransaction'
        "500":
          description: Failed to retrieve transaction data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Find all
      tags:
      - Transaction Query
  /api/transaction-query/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a transaction record using its ID
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Transaction data
          schema:
            $ref: '#/definitions/response.ApiResponseTransaction'
        "400":
          description: 'Bad Request: Invalid ID'
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to retrieve transaction data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Find a transaction by ID
      tags:
      - Transaction Query
  /api/transaction-query/active:
    get:
      consumes:
      - application/json
      description: Retrieve a list of active transactions
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Number of items per page (default: 10)'
        in: query
        name: page_size
        type: integer
      - description: Search keyword
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of active transactions
          schema:
            $ref: '#/definitions/response.ApiResponseTransactions'
        "500":
          description: Failed to retrieve transaction data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Find active transactions
      tags:
      - Transaction Query
  /api/transaction-query/card-number/{card_number}:
    get:
      consumes:
      - application/json
      description: Retrieve a list of transactions for a specific card number
      parameters:
      - description: Card Number
        in: path
        name: card_number
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of transactions
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationTransaction'
        "500":
          description: Failed to retrieve transaction data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Find all transactions by card number
      tags:
      - Transaction Query
  /api/transaction-query/trashed:
    get:
      consumes:
      - application/json
      description: Retrieve a list of trashed transactions
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Number of items per page (default: 10)'
        in: query
        name: page_size
        type: integer
      - description: Search keyword
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of trashed transactions
          schema:
            $ref: '#/definitions/response.ApiResponseTransactions'
        "500":
          description: Failed to retrieve transaction data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Retrieve trashed transactions
      tags:
      - Transaction Query
...
  /api/transactions-query/merchant/{merchant_id}:
    get:
      consumes:
      - application/json
      description: Retrieve a list of transactions using the merchant ID
      parameters:
      - description: Merchant ID
        in: query
        name: merchant_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Transaction data
          schema:
            $ref: '#/definitions/response.ApiResponseTransactions'
        "400":
          description: 'Bad Request: Invalid ID'
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to retrieve transaction data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Find transactions by merchant ID
      tags:
      - Transaction
...
swagger: "2.0"